+++ test/runtests.jl
@@ -48,6 +48,7 @@ include("optsummary.jl")
 include("predict.jl")
 include("sigma.jl")
 include("grad.jl")
+include("RFP.jl")
 
 @testset "PRIMA" include("prima.jl")
 @testset "ForwardDiff" include("forwarddiff.jl")